\name{basis.toy}
\alias{basis.toy}
\title{Toy basis functions}
\description{
A working example of a basis function
}
\usage{
basis.toy(x)
}
\arguments{
  \item{x}{Point in parameter space}
}
\references{
  M. C. Kennedy and A. O'Hagan 2000. \dQuote{Predicting the output from
    a complex computer code when fast approximations are available}
  Biometrika, 87(1): pp1-13
  }
\author{Robin K. S. Hankin}
\examples{
data(toyapps)
basis.toy(D1.toy)
}
\keyword{array}
