\name{betas.toy}
\alias{betas.toy}
\docType{data}
\title{Toy example of coefficients beta}
\description{
Toy example of coefficients beta.
}
\usage{data(betas.toy)}
\format{
This is a matrix with four rows (one per level) and five columns (one
per term of \code{basis.toy})
}
\examples{
data(betas.toy)
}
\keyword{datasets}
